Added the first component
It works poorly. Next is to add the results and preview components and hopefully then just turn on the backend and it should work (maybe need to change port on the server or on next.js). Also error handling is not existing anymore it seems, so need to add that somehow. It does not write out failed to fetch on the screen anymore but some error logs instead. Also it renders a spinner so that's good, but need to make the spinner reusable everywhere so it should not stay in the page.js like current.